Separate tracing examples into distinct projects
