+++ projects/xrpl_std_examples/trace_escrow_account/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "xrpl_std_example"
+name = "trace_escrow_account"
 version = "0.0.1"
 edition = "2024"
-description = "A Smart Escrow that calls every host function and accesses every field of every Craft primitive."
+description = "A Smart Escrow that accesses and traces every field in an `AccountRoot` object associated with an `EscrowFinish`"
 license = "MIT"
 
 # This empty workspace definition keeps this project independent of the parent workspace
@@ -21,4 +21,4 @@ panic = "abort"   # `no-std` can't unwind on panic, so choose abort (plus no pan
 panic = "unwind" # For debugging, support std-lib with unwinding in the `dev` profile
 
 [dependencies]
-xrpl-std = { path = "../../xrpl-std" }
+xrpl-std = { path = "../../../xrpl-std" }
