+++ projects/xrpl_std_examples/trace_escrow_account/fixtures/success/ledger_object.json
