+++ projects/xrpl_std_examples/trace_escrow_account/fixtures/success/tx.json
@@ -1,0 +1,46 @@
+{
+  "Account": "rHb9CJAWyB4rj91VRWn96DkukG4bwdtyTh",
+  "TransactionType": "EscrowFinish",
+  "ComputationAllowance": "1000001",
+  "Fee": "10",
+  "Sequence": 4294967295,
+  "Flags": 4294967294,
+  "OfferSequence": 4294967293,
+  "Owner": "rrrrrrrrrrrrrrrrrrrrrhoLvTp",
+  "SigningPubKey": "0330E7FC9D56BB25D6893BA3F317AE5BCF33B3291BD63DB32654A313222F7FD020",
+  "TxnSignature": "30450221008AD5EE48F7F1047813E79C174FE401D023A4B4A7B99AF826E081DB1DFF7B9C510220133F05B7FD3D7D7F163E8C77EE0A49D02619AB6C77CC3487D0095C9B34033C1C",
+  "hash": "74465121372813CBA4C77E31F12E137163F5B2509B16AC1703ECF0DA194B2DD4",
+  "AccountTxnID": "DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD",
+  "LastLedgerSequence": 4294967292,
+  "NetworkID": 4294967291,
+  "SourceTag": 4294967290,
+  "TicketSequence": 4294967289,
+  "Condition": "3333333333333333333333333333333333333333333333333333333333333333",
+  "Fulfillment": "2121212121212121212121212121212121212121212121212121212121212121",
+  "CredentialIDs": [
+    "0ABA05A34949F2CED41025914FC4F267883F1D388A6545AFB4863466FAA6F28C",
+    "D0A063DEE0B0EC9522CF35CD55771B5DCAFA19A133EE46A0295E4D089AF86438",
+    "D2EFD38589609AE570D17E9957CE6002E764A63EE66FE8CAA2768976ABD60BFF"
+  ],
+  "Memos": [
+    {
+      "Memo": {
+        "MemoType": "687474703a2f2f6578616d706c652e636f6d2f6d656d6f2f67656e65726963",
+        "MemoData": "72656e74",
+        "MemoFormat": "687474703a2f2f6578616d706c652e636f6d2f6d656d6f2f67656e65726963"
+      }
+    }
+  ],
+  "Signers": [
+    {
+      "Account": "rsA2LpzuawewSBQXkiju3YQTMzW13pAAdW",
+      "TxnSignature": "3333333333333333333333333333333333333333333333333333333333333333",
+      "SigningPubKey": "0330E7FC9D56BB25D6893BA3F317AE5BCF33B3291BD63DB32654A313222F7FD020"
+    },
+    {
+      "Account": "rrrrrrrrrrrrrrrrrrrrrhoLvTp",
+      "TxnSignature": "3333333333333333333333333333333333333333333333333333333333333333",
+      "SigningPubKey": "0330E7FC9D56BB25D6893BA3F317AE5BCF33B3291BD63DB32654A313222F7FD020"
+    }
+  ]
+}
