+++ projects/xrpl_std_examples/trace_escrow_finish/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "trace_escrow_finish"
+version = "0.0.1"
+edition = "2024"
+description = "A Smart Escrow that accesses and traces every field in the `EscrowFinish` transaction of a Smart Escrow."
+license = "MIT"
+
+# This empty workspace definition keeps this project independent of the parent workspace
+[workspace]
+
+[lib]
+crate-type = ["cdylib"]
+
+[profile.release]
+opt-level = "s"   # Optimize for size
+lto = true        # Link-time optimization
+codegen-units = 1 # Compile in one unit for better optimization and smaller binary size
+panic = "abort"   # `no-std` can't unwind on panic, so choose abort (plus no panic handler means smaller binary)
+
+[profile.dev]
+panic = "unwind" # For debugging, support std-lib with unwinding in the `dev` profile
+
+[dependencies]
+xrpl-std = { path = "../../../xrpl-std" }
