+++ projects/xrpl_std_examples/trace_escrow_finish/fixtures/success/ledger_object.json
@@ -1,0 +1,3 @@
+{
+
+}
