+++ projects/xrpl_std_examples/trace_escrow_finish/fixtures/success/nfts.json
@@ -1,0 +1,3 @@
+[
+
+]
