+++ projects/xrpl_std_examples/trace_escrow_ledger_object/README.md
@@ -1,0 +1,39 @@
+# XRPL Std Lib Example
+
+This WebAssembly module demonstrates how to use the XRPL std lib to access every field in an `Escrow` ledger object that
+is being finished by a Smart Escrow contract. The WASM program is meant to both execute known host functions required
+for accessing the fields of an `Escrow` ledger object, and also validate that field access is working correctly by
+asserting each value that craft makes available. In this way, this Smart Escrow can be used as a type of canary that can
+indicate if anything in Craft is not operating according to expectations.
+
+## Building
+
+Build using:
+
+```bash
+# Navigate to the project directory
+cd projects/xrpl_std_examples/trace_escrow_ledger_object
+
+# Build the WASM file
+cargo build --target wasm32-unknown-unknown --release
+```
+
+The resulting WASM file will be located at:
+
+```
+./target/wasm32-unknown-unknown/release/trace_escrow_ledger_object.wasm
+```
+
+## Running with wasm-host
+
+Run the contract using the wasm-host application:
+
+```bash
+cd ../../../wasm-host
+cargo run -- --wasm-file ../projects/xrpl_std_examples/trace_escrow_ledger_object/target/wasm32-unknown-unknown/release/trace_escrow_ledger_object.wasm --project xrpl_std_examples/trace_escrow_ledger_object
+```
+
+### Note
+
+Please note that the wasm-host only has mock host functions. Please use the devnet (or a standalone rippled node) to
+test with a real implementation and real data.
