+++ projects/xrpl_std_examples/trace_escrow_ledger_object/fixtures/success/ledger.json
@@ -1,0 +1,1 @@
+[]
