+++ projects/xrpl_std_examples/trace_escrow_ledger_object/fixtures/success/ledger_object.json
@@ -1,0 +1,17 @@
+{
+  "Account": "rf1BiGeXwwQoi8Z2ueFYTEXSwuJYfV2Jpn",
+  "Amount": "10000",
+  "CancelAfter": 545440232,
+  "Condition": "A0258020A82A88B2DF843A54F58772E4A3861866ECDB4157645DD9AE528C1D3A",
+  "Destination": "ra5nK24KXen9AHvsdFTKHSANinZseWnPcX",
+  "DestinationTag": 23480,
+  "FinishAfter": 545354132,
+  "Flags": 0,
+  "FinishFunction": "0061736D0100000001180460027F7F0060017F017F60027F7F017F60047F7F7F7F00030C0B01010200000000000003000405017001030305030100110619037F01418080C0000B7F0041DD85C0000B7F0041E085C0000B074205066D656D6F7279020008616C6C6F6361746500000F636865636B5F6163636F756E74494400020A5F5F646174615F656E6403010B5F5F686561705F6261736503020908010041010B02060A0AF5360B610002",
+  "LedgerEntryType": "Escrow",
+  "OwnerNode": "0000000000000000",
+  "DestinationNode": "0000000000000000",
+  "PreviousTxnID": "C44F2EB84196B9AD820313DBEBA6316A15C9A2D35787579ED172B87A30131DA7",
+  "PreviousTxnLgrSeq": 28991004,
+  "SourceTag": 11747
+}
